requirements.txt
beautifulsoup4
flask
Flask-Cache
Flask-SQLAlchemy
psycopg2
redis
requests
